+++ CHANGELOG.md
@@ -12,7 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - SSAO resolution scale (#1423, **@tomas7770**).
 - Warm starting for collision solving (#1248, **@fallenatlas**).
 - Conditional shader compilation (#1406, **@tomas7770**).
-- Collision layers and masks, for filtering collisions (**@RiscadoA**).
+- Collision layers and masks, for filtering collisions and raycasts (**@RiscadoA**).
 
 ### Changed
 
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Normalize normals while drawing to the GBuffer to avoid shiny small objects (**@RiscadoA**).
 - Make FreeCameraController respect other plugins locking the mouse (**@RiscadoA**).
 - Draw Transform Gizmos for all active cameras instead of only the first one (**@RiscadoA**).
+- Consider Collider component transform matrix while doing Raycast (**@RiscadoA**)
 
 ## [v0.5.0] - 2024-12-01